refactor(nn): moved mac operators to vhdl shared design
