+++ elasticai/creator/vhdl/shared_designs/mac/binary/bin_mac.vhd
