+++ elasticai/creator/vhdl/shared_designs/mac/binary/testbench.tpl.vhd
