+++ elasticai/creator/vhdl/shared_designs/mac/fixed_point/fxp_mac.tpl.vhd
