+++ elasticai/creator/vhdl/shared_designs/mac/fixed_point/mac_design.tpl.vhd
